Sometimes, streams just never end
Previously, this relied on reaching the end of a stream before emitting concatenated angular modules.
In a situation where the stream is piping from a watcher, this meant the stream was never closed.
Consequently, the stream was never flushed, and modules were never emitted!

Instead, do emit events as files are encountered and a module definition has already happened.
Also, support updating the file and re-emitting concatenated module if an existing file is encountered again.